Integrate truss module into main simulator
